requirements.txt
dimod==0.12.3
dwave_samplers==1.0.0
numpy==1.21.2
pandas==1.3.0
scipy==1.9.1
sympy==1.10.dev0
